pyproject.toml
[tool.poetry]
name = "kurs4"
version = "0.1.0"
description = ""
authors = ["danilu <danilu2537@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.19"
flask = "^2.3.2"
flask-cors = "^4.0.0"
flask-restx = "^1.1.0"
flask-sqlalchemy = "^3.0.5"
marshmallow = "^3.20.1"
pyjwt = "^2.8.0"
pytest = "^7.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
skip-string-normalization = true
skip-magic-trailing-comma = true
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | migrations
)/
'''

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["B", "E", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    'migrations',
    '.mypy_cache',
]

line-length = 88
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 88
combine_as_imports = true
combine_star = true
skip_glob = '''
  .git,
  .mypy_cache,
  .venv,
  migrations
'''
